moved progress_bar function from hek2vso module and created a module in the sunpy.util directory
