+++ sunpy/net/hek2vso/hek2vso.py
@@ -17,6 +17,7 @@ import sys
 from astropy import units
 from sunpy.net import hek
 from sunpy.net import vso
+from sunpy.util import progressbar2 as PB
 
 __author__ = 'Michael Malocha'
 __version__ = 'Aug 10th, 2013'
@@ -134,46 +135,6 @@ def vso_attribute_parse(phrase):
     return query
 
 
-def progress_bar(phrase, position, total, bar_size=20):
-    """
-    Prints a simple progress bar to the screen
-
-    Parameters
-    ----------
-    phrase: str
-        The desired phrase to precede the progress bar with each update.
-    position: int
-        The current location in the data set (will be divided by 'total'
-        to determine percent finished).
-    total: int
-        The total size of the data set
-    bar_size: int
-        Size of the bar, defaults to 20 characters in length
-
-    Examples
-    --------
-    >>> progress_bar("Progress:", 10, 100)
-    Progress: [##                  ] 10%
-
-    >>> progress_bar("Progress:", 35, 100)
-    Progress: [#######             ] 35%
-
-    >>> progress_bar("Progress:", 35, 83)
-    Progress: [########            ] 42%
-    """
-    position = float(format(position, '.1f'))
-    fraction = position / total
-    percent = str(int(100 * fraction)) + '%'
-    place = int(fraction * bar_size)
-    pounds = '#' * place
-    blanks = ' ' * (20 - place)
-    prog_bar = ' [' + pounds + blanks + '] ' + percent
-    sys.stdout.write('\r' + ' ' * 52)
-    sys.stdout.flush()
-    sys.stdout.write('\r' + phrase + prog_bar)
-    sys.stdout.flush()
-
-
 class H2VClient(object):
     """
     Class to handle HEK to VSO translations
@@ -268,7 +229,7 @@ class H2VClient(object):
         place = 1
         for query in vso_query:
             if use_progress_bar:
-                progress_bar('Querying VSO webservice', place, result_size)
+                PB.progress_bar('Querying VSO webservice', place, result_size)
             temp = self.vso_client.query(*query)
             self.vso_results.append(temp)
             self.num_of_records += len(temp)
